UIQM-716: Consolidate routes based on MARC type for bib and authority records to avoid page refresh after redirecting from the create page to the edit one.
