+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@folio/quick-marc",
-  "version": "9.0.0",
+  "version": "10.0.0",
   "description": "Quick MARC editor",
   "main": "index.js",
   "repository": "",
