+++ src/QuickMarc.js
@@ -2,16 +2,12 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import {
   Switch,
+  useLocation,
 } from 'react-router-dom';
 
 import { CommandList } from '@folio/stripes/components';
 
 import { MarcRoute } from './MarcRoute';
-import {
-  QuickMarcDeriveWrapper,
-  QuickMarcCreateWrapper,
-  QuickMarcEditWrapper,
-} from './QuickMarcEditor';
 import { QUICK_MARC_ACTIONS } from './QuickMarcEditor/constants';
 import {
   MARC_TYPES,
@@ -24,32 +20,34 @@ const QuickMarc = ({
   onClose,
   onSave,
 }) => {
+  const location = useLocation();
+
+  const permissionsMap = {
+    'create-bibliographic': 'ui-quick-marc.quick-marc-editor.create',
+    'edit-bibliographic': 'ui-quick-marc.quick-marc-editor.all',
+    'derive-bibliographic': 'ui-quick-marc.quick-marc-editor.derive.execute',
+    'create-authority': 'ui-quick-marc.quick-marc-authorities-editor.create',
+    'edit-authority': '', // ui-quick-marc.quick-marc-authorities-editor.all
+  };
+
+  // .../some-path/create-bibliographic => [, create-bibliographic, create, bibliographic]
+  const [, page, action] = location.pathname.match(/\/((edit|create|derive)-(bibliographic|authority|holdings))/) || [];
+
   const editorRoutesConfig = [
     {
-      path: `${basePath}/edit-bib/:externalId`,
-      permission: 'ui-quick-marc.quick-marc-editor.all',
+      path: `${basePath}/:action-bibliographic/:externalId?`,
+      permission: permissionsMap[page],
       props: {
-        action: QUICK_MARC_ACTIONS.EDIT,
-        wrapper: QuickMarcEditWrapper,
+        action,
         marcType: MARC_TYPES.BIB,
       },
     },
     {
-      path: `${basePath}/duplicate-bib/:externalId`,
-      permission: 'ui-quick-marc.quick-marc-editor.derive.execute',
+      path: `${basePath}/:action-authority/:externalId?`,
+      permission: permissionsMap[page],
       props: {
-        action: QUICK_MARC_ACTIONS.DERIVE,
-        wrapper: QuickMarcDeriveWrapper,
-        marcType: MARC_TYPES.BIB,
-      },
-    },
-    {
-      path: `${basePath}/create-bib`,
-      permission: 'ui-quick-marc.quick-marc-editor.create',
-      props: {
-        action: QUICK_MARC_ACTIONS.CREATE,
-        wrapper: QuickMarcCreateWrapper,
-        marcType: MARC_TYPES.BIB,
+        action,
+        marcType: MARC_TYPES.AUTHORITY,
       },
     },
     {
@@ -57,7 +55,6 @@ const QuickMarc = ({
       permission: 'ui-quick-marc.quick-marc-holdings-editor.create',
       props: {
         action: QUICK_MARC_ACTIONS.CREATE,
-        wrapper: QuickMarcCreateWrapper,
         marcType: MARC_TYPES.HOLDINGS,
       },
     },
@@ -66,26 +63,7 @@ const QuickMarc = ({
       permission: 'ui-quick-marc.quick-marc-holdings-editor.all',
       props: {
         action: QUICK_MARC_ACTIONS.EDIT,
-        wrapper: QuickMarcEditWrapper,
         marcType: MARC_TYPES.HOLDINGS,
-      },
-    },
-    {
-      path: `${basePath}/create-authority`,
-      permission: 'ui-quick-marc.quick-marc-authorities-editor.create',
-      props: {
-        action: QUICK_MARC_ACTIONS.CREATE,
-        wrapper: QuickMarcCreateWrapper,
-        marcType: MARC_TYPES.AUTHORITY,
-      },
-    },
-    {
-      path: `${basePath}/edit-authority/:externalId`,
-      // permission: 'ui-quick-marc.quick-marc-authorities-editor.all',
-      props: {
-        action: QUICK_MARC_ACTIONS.EDIT,
-        wrapper: QuickMarcEditWrapper,
-        marcType: MARC_TYPES.AUTHORITY,
       },
     },
   ];
@@ -108,6 +86,7 @@ const QuickMarc = ({
                 path={path}
                 permission={permission}
                 routeProps={routeProps}
+                basePath={basePath}
                 onClose={onClose}
                 onSave={onSave}
               />
