+++ src/QuickMarcEditor/index.js
@@ -1,5 +1,2 @@
 export { default as QuickMarcEditor } from './QuickMarcEditor';
 export { default as QuickMarcEditorContainer } from './QuickMarcEditorContainer';
-export { default as QuickMarcDeriveWrapper } from './QuickMarcDeriveWrapper';
-export { default as QuickMarcCreateWrapper } from './QuickMarcCreateWrapper';
-export { default as QuickMarcEditWrapper } from './QuickMarcEditWrapper';
