+++ src/QuickMarcEditor/useSaveRecord/useSubmitRecord/index.js
@@ -1,0 +1,1 @@
+export * from './useSumbitRecord';
